define webhook url prefix constant
