+++ src/main/java/io/jenkins/plugins/sdk/DingTalkRobotRequest.java
@@ -1,5 +1,6 @@
 package io.jenkins.plugins.sdk;
 
+import io.jenkins.plugins.tools.Constants;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
