+++ src/main/java/io/jenkins/plugins/sdk/DingTalkRobotResponse.java
@@ -4,13 +4,11 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * @author liuwei
+ * DingTalkRobotResponse <a href="https://open.dingtalk.com/document/orgapp/custom-robots-send-group-messages#h2-zl3-3lb-eww">...</a>
  */
 @Data
 @NoArgsConstructor
 public class DingTalkRobotResponse {
-
-private int errcode;
-
-private String errmsg;
+    private int errcode;
+    private String errmsg;
 }
